+++ code/src/components/Header.js
@@ -1,10 +1,10 @@
 import React from "react";
-import moment from 'moment';
+import moment from "moment";
 
 import "../css/header.css";
 
 export const Header = () => {
-  const todaysDate = moment().format("YYYY-MM-DD")
+  const todaysDate = moment().format("YYYY-MM-DD");
   return (
     <header>
       <h1>Todo</h1>
