+++ code/src/css/addTodo.css
@@ -18,7 +18,6 @@
 .add-todo-button {
   cursor: pointer;
   width: 55px;
-  /* background: linear-gradient(#6c7a89, #333); */
   background: linear-gradient(#5c7c99, #395873, #063053);
   color: #e5e5e5;
   border: none;
@@ -26,8 +25,8 @@
 }
 
 .add-todo-button:disabled {
-    cursor: auto;
-    opacity: 0.8;
+  cursor: auto;
+  opacity: 0.8;
 }
 
 .add-todo-button:enabled:active {
