+++ code/src/css/statusBar.css
@@ -1,5 +1,4 @@
 .status-bar {
-  /* background: linear-gradient(#6c7a89, #a6afb8); */
   background: #5c7c99;
   color: #eee;
   padding: 10px 10px 10px 15px;
@@ -11,47 +10,3 @@
   margin-top: 15px;
 }
 
-.button-container {
-  display: flex;
-  flex-direction: column;  
-}
-
-.clear-complete-button {
-  margin-bottom: 5px;
-}
-
-.clear-all-button,
-.clear-complete-button {
-  cursor: pointer;
-  padding: 10px;
-  margin-left: 10px;
-  border-radius: 12px;
-  border: none;
-  text-transform: uppercase;
-  font-size: 12px;
-  background: linear-gradient(#395873, #063053);
-  color: #eee;
-  outline: none;
-  font-weight: 400;
-  transition: letter-spacing 500ms ease;
-}
-
-.clear-all-button:hover,
-.clear-complete-button:hover {
-  color: #fff;
-  letter-spacing: 0.7px;
-}
-
-.clear-all-button:active,
-.clear-complete-button:active {
-  transform: translateY(1px);
-}
-
-@media (min-width: 500px) {
-  .button-container {
-    flex-direction: row;  
-  }
-  .clear-complete-button {
-    margin-bottom: 0;
-  }
-}
