+++ code/src/css/todoList.css
@@ -4,68 +4,3 @@
   align-items: center;
   width: 100%;
 }
-
-.todo-item {
-  padding-left: 15px;
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  background: #eee;
-  color: #063053;
-  margin-bottom: 15px;
-  border: 1px solid #bebebe;
-  box-shadow: 3px 5px 8px 1px rgba(0, 0, 0, 0.2);
-  transition: box-shadow 500ms ease;
-}
-
-.container {
-    display: flex;
-    align-items: center;
-}
-
-.todo-item:hover {
-  box-shadow: 3px 8px 16px 1px rgba(0, 0, 0, 0.3);
-}
-
-.complete-checkbox {
-  width: 25px;
-  height: 25px;
-  cursor: pointer;
-  margin-right: 15px;
-}
-
-.complete-checkbox input {
-  display: none;
-}
-
-.checkmark {
-  width: 100%;
-  height: 100%;
-  border: 1px solid #063053;
-  border-radius: 3px;
-  display: inline-block;
-  background: #5c7c99 url("../img/checked.png") center/0 no-repeat;
-  transition: background-size 300ms ease-in;
-}
-
-.complete-checkbox input:checked + .checkmark {
-  background-size: 60%;
-}
-
-.text-when-checked {
-    text-decoration: line-through;
-  }
-
-.remove-icon {
-  margin-right: -5px;
-  transform: scale(0.4);
-  transition: transform 300ms ease-in-out;
-  outline: none;
-}
-
-.remove-icon:hover {
-  transform: scale(0.5);
-}
-
-
